refactor(probe): migrate igmp to per-protocol plugin
