+++ tools/probe/engine/cli.py
@@ -1317,19 +1317,10 @@ def _failure_reasons_for_case(case_name: str) -> list[str]:
     # plugin's ``failure_reasons`` hook
     # (:func:`tools.probe.engine.protocols.udp.udp_failure_reasons`), dispatched
     # registry-first above, so it no longer appears in this legacy chain.
-    if case_name in {
-        "igmp-membership-query-observation",
-        "igmp-v2-membership-report-emission",
-        "igmp-v2-leave-group-emission",
-        "igmp-v3-source-list-report",
-    }:
-        return [
-            FAILURE_TIMEOUT,
-            FAILURE_WRONG_PEER,
-            FAILURE_WRONG_PAYLOAD,
-            FAILURE_DECODE_FAILED,
-            FAILURE_TARGET_SETUP_FAILED,
-        ]
+    # The IGMP failure-reason taxonomy (the four IGMP cases) moved to the IGMP
+    # plugin's ``failure_reasons`` hook
+    # (:func:`tools.probe.engine.protocols.igmp.igmp_failure_reasons`), dispatched
+    # registry-first above, so it no longer appears in this legacy chain.
     return _default_failure_reasons()
