+++ tools/probe/engine/planning.py
@@ -196,6 +196,19 @@ from .protocols.rip import (  # noqa: F401  (re-exported for identity/back-compa
     _rip_update_probe_plan,
     _ripng_update_probe_plan,
 )
+# The IGMP planning surface (the four IGMP cases' shared ``_igmp_probe_plan``
+# builder, the ``_igmp_target_service`` plan-building helper, and the
+# ``deterministic_igmp_group`` / ``deterministic_igmp_source_list`` helpers) lives
+# in the IGMP plugin module. Re-import the moved builder and deterministic helpers
+# so ``planning._igmp_probe_plan`` / ``planning.deterministic_igmp_group`` /
+# ``planning.deterministic_igmp_source_list`` resolve to the *same* function
+# objects the plugin registered and the merged ``PLAN_BUILDERS`` exposes -- any
+# pin on object identity keeps holding.
+from .protocols.igmp import (  # noqa: F401  (re-exported for identity/back-compat)
+    _igmp_probe_plan,
+    deterministic_igmp_group,
+    deterministic_igmp_source_list,
+)
 
 
 def planned_cases(
@@ -293,307 +306,19 @@ _IKEV2_UDP_PORT = 500
 # are re-imported into this module above for identity/back-compat.
 
 
-# IGMP (IPv4 protocol number 2) rides link-local IPv4 multicast with TTL 1 and
-# Router Alert on real networks. The probe plans below are dry-run-only and
-# provider-backed for live execution; they expose the exact message shapes an
-# agent should expect before any protected live run is confirmed.
-_IGMP_IP_PROTOCOL = 2
-_IGMP_DEFAULT_TTL = 1
-_IGMP_ALL_SYSTEMS_GROUP = "224.0.0.1"
-_IGMP_ALL_ROUTERS_GROUP = "224.0.0.2"
-_IGMPV3_REPORT_DESTINATION = "224.0.0.22"
-_IGMP_DOCUMENTATION_MULTICAST_PREFIX = "233.252.0.0/24"
-_IGMP_DOCUMENTATION_SOURCE_PREFIXES = ["192.0.2.0/24", "198.51.100.0/24"]
-_IGMP_TARGET_SERVICE_DIR = "tools/probe/target_services/igmp"
-_IGMP_LISTENER_SCRIPT = f"{_IGMP_TARGET_SERVICE_DIR}/provision-listener.sh"
-_IGMP_ROUTER_SCRIPT = f"{_IGMP_TARGET_SERVICE_DIR}/provision-router.sh"
-_IGMP_CLEANUP_SCRIPT = f"{_IGMP_TARGET_SERVICE_DIR}/cleanup.sh"
+# The IGMP planning surface (the four IGMP cases' shared ``_igmp_probe_plan``
+# builder, the ``_igmp_target_service`` plan-building helper, the
+# ``deterministic_igmp_group`` / ``deterministic_igmp_source_list`` helpers, and
+# the IGMP wire constants they reference) now lives in the IGMP plugin module
+# (``protocols/igmp.py``); the moved builder reaches ``PLAN_BUILDERS`` through the
+# registry merge below and is re-imported into this module above for
+# identity/back-compat.
 
 
 # The BGP plan builder (``_bgp_session_smoke_probe_plan``) now lives in the BGP
 # plugin module (``protocols/bgp.py``) and reaches ``PLAN_BUILDERS`` through the
 # registry merge below; it is re-imported into this module above for
 # identity/back-compat.
-
-
-def _igmp_probe_plan(
-    *,
-    case_name: str,
-    profile: str,
-    seed: int,
-    sequence: int,
-) -> JSONObject:
-    """Plan a dry-run IGMP peer-behavior case.
-
-    IGMP probes need a provider-backed multicast segment and a controlled peer
-    (listener/router) before any live traffic is useful. This builder therefore
-    records packet shapes, capture filters, target-service scripts, and stable
-    validation expectations while emitting no packet bytes and sending nothing.
-    """
-
-    case = PROBE_CASE_BY_NAME[case_name]
-    digest = deterministic_bytes(case_name, profile, seed, sequence)
-    stimulus_ipv4, target_ipv4 = deterministic_ipv4_pair(profile, seed, sequence)
-    group_address = deterministic_igmp_group(profile, seed, sequence)
-    source_list = deterministic_igmp_source_list(profile, seed, sequence)
-
-    if case_name == "igmp-membership-query-observation":
-        destination_ipv4 = _IGMP_ALL_SYSTEMS_GROUP
-        sender_ipv4 = target_ipv4
-        target_service = _igmp_target_service(
-            kind="igmp-router",
-            role="router",
-            bind_ipv4=target_ipv4,
-            source_ipv4=stimulus_ipv4,
-            multicast_group=destination_ipv4,
-            provision_script=_IGMP_ROUTER_SCRIPT,
-        )
-        stimulus_shape: JSONObject = {
-            "direction": "observe",
-            "sender_role": "router",
-            "igmp_type": 0x11,
-            "message": "membership_query",
-            "max_response_time_tenths": 10,
-            "group_address": "0.0.0.0",
-        }
-        expected_shape = dict(stimulus_shape)
-        expected_shape["source_ipv4"] = sender_ipv4
-        expected_shape["destination_ipv4"] = destination_ipv4
-        validation: JSONObject = {
-            "planned_only": True,
-            "driver": "igmp_query_observation",
-            "source_ipv4": sender_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x11,
-            "group_address": "0.0.0.0",
-            "max_response_time_tenths": 10,
-        }
-        capture_filter = (
-            f"igmp and src host {sender_ipv4} and dst host {destination_ipv4}"
-        )
-        adapter_case = "igmp-v2-membership-query"
-    elif case_name == "igmp-v2-membership-report-emission":
-        destination_ipv4 = group_address
-        target_service = _igmp_target_service(
-            kind="igmp-listener",
-            role="target",
-            bind_ipv4=target_ipv4,
-            source_ipv4=stimulus_ipv4,
-            multicast_group=group_address,
-            provision_script=_IGMP_LISTENER_SCRIPT,
-        )
-        stimulus_shape = {
-            "direction": "emit",
-            "igmp_type": 0x16,
-            "message": "v2_membership_report",
-            "group_address": group_address,
-        }
-        expected_shape = {
-            "observer_role": "target",
-            "expected_response": case.expected_response,
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x16,
-            "group_address": group_address,
-        }
-        validation = {
-            "planned_only": True,
-            "driver": "igmp_v2_membership_report",
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x16,
-            "group_address": group_address,
-        }
-        capture_filter = (
-            f"igmp and src host {stimulus_ipv4} and dst host {destination_ipv4}"
-        )
-        adapter_case = "igmp-v2-membership-report"
-    elif case_name == "igmp-v2-leave-group-emission":
-        destination_ipv4 = _IGMP_ALL_ROUTERS_GROUP
-        target_service = _igmp_target_service(
-            kind="igmp-listener",
-            role="target",
-            bind_ipv4=target_ipv4,
-            source_ipv4=stimulus_ipv4,
-            multicast_group=group_address,
-            provision_script=_IGMP_LISTENER_SCRIPT,
-        )
-        stimulus_shape = {
-            "direction": "emit",
-            "igmp_type": 0x17,
-            "message": "v2_leave_group",
-            "group_address": group_address,
-        }
-        expected_shape = {
-            "observer_role": "target",
-            "expected_response": case.expected_response,
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x17,
-            "group_address": group_address,
-        }
-        validation = {
-            "planned_only": True,
-            "driver": "igmp_v2_leave_group",
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x17,
-            "group_address": group_address,
-        }
-        capture_filter = (
-            f"igmp and src host {stimulus_ipv4} and dst host {destination_ipv4}"
-        )
-        adapter_case = "igmp-v2-leave-group"
-    elif case_name == "igmp-v3-source-list-report":
-        destination_ipv4 = _IGMPV3_REPORT_DESTINATION
-        target_service = _igmp_target_service(
-            kind="igmp-listener",
-            role="target",
-            bind_ipv4=target_ipv4,
-            source_ipv4=stimulus_ipv4,
-            multicast_group=group_address,
-            provision_script=_IGMP_LISTENER_SCRIPT,
-        )
-        stimulus_shape = {
-            "direction": "emit",
-            "igmp_type": 0x22,
-            "message": "v3_membership_report",
-            "record_type": "mode_is_include",
-            "group_address": group_address,
-            "source_addresses": source_list,
-            "record_count": 1,
-        }
-        expected_shape = {
-            "observer_role": "target",
-            "expected_response": case.expected_response,
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x22,
-            "group_address": group_address,
-            "source_addresses": source_list,
-            "record_type": "mode_is_include",
-        }
-        validation = {
-            "planned_only": True,
-            "driver": "igmp_v3_source_list_report",
-            "source_ipv4": stimulus_ipv4,
-            "destination_ipv4": destination_ipv4,
-            "igmp_type": 0x22,
-            "group_address": group_address,
-            "source_addresses": source_list,
-            "record_type": "mode_is_include",
-            "record_count": 1,
-        }
-        capture_filter = (
-            f"igmp and src host {stimulus_ipv4} and dst host {destination_ipv4}"
-        )
-        adapter_case = "igmp-v3-source-list-report"
-    else:
-        raise ValueError(f"unsupported IGMP probe case {case_name!r}")
-
-    return {
-        "schema_version": 1,
-        "case": case_name,
-        "sequence": sequence,
-        "index": sequence,
-        "profile": profile,
-        "seed": seed,
-        "stimulus": case.stimulus,
-        "expected_response": case.expected_response,
-        "planned_only": True,
-        "protocol": "igmp",
-        "ip_protocol": _IGMP_IP_PROTOCOL,
-        "source_ipv4": stimulus_ipv4,
-        "destination_ipv4": destination_ipv4,
-        "expected_reply_source_ipv4": target_ipv4,
-        "expected_reply_destination_ipv4": stimulus_ipv4,
-        "group_address": group_address,
-        "multicast_group": group_address,
-        "ttl": _IGMP_DEFAULT_TTL,
-        "router_alert_required": True,
-        "documentation_prefixes": [
-            _IGMP_DOCUMENTATION_MULTICAST_PREFIX,
-            *_IGMP_DOCUMENTATION_SOURCE_PREFIXES,
-        ],
-        "stimulus_driver": {
-            "name": case.stimulus,
-            "adapter_module": "tools/probe/adapters/src/igmp.rs",
-            "adapter_case": adapter_case,
-            "state": "planned-only",
-            "planned_only": True,
-        },
-        "stimulus_packet_shape": {
-            "ipv4": {
-                "source": stimulus_ipv4,
-                "destination": destination_ipv4,
-                "ttl": _IGMP_DEFAULT_TTL,
-                "protocol": _IGMP_IP_PROTOCOL,
-                "router_alert_required": True,
-            },
-            "igmp": stimulus_shape,
-        },
-        "expected_response_packet_shape": expected_shape,
-        "target_service": target_service,
-        "capture_filter": capture_filter,
-        "validation": validation,
-        "wire_requirements": {
-            "requires_ipv4_multicast": True,
-            "requires_igmp_peer": True,
-            "requires_provider_backing": True,
-            "requires_confirm_live_run": True,
-            "note": (
-                "IGMP dry-run records the packet and peer-observation contract. "
-                "Any live traffic must run from a provider-backed lab endpoint "
-                "with explicit confirmation, never from the developer host."
-            ),
-        },
-        "live_path": (
-            "Opt-in via lab-session / providers: provision the IGMP target "
-            "listener/router, run the stimulus from the provider endpoint with "
-            "--confirm-live-run, collect artifacts, and tear the endpoint down."
-        ),
-        "digest_hex": digest.hex()[:16],
-    }
-
-
-def _igmp_target_service(
-    *,
-    kind: str,
-    role: str,
-    bind_ipv4: str,
-    source_ipv4: str,
-    multicast_group: str,
-    provision_script: str,
-) -> JSONObject:
-    return {
-        "required": True,
-        "kind": kind,
-        "protocol": "igmp",
-        "role": role,
-        "bind_ipv4": bind_ipv4,
-        "source_ipv4": source_ipv4,
-        "multicast_group": multicast_group,
-        "runtime": "probe-target-service",
-        "provision_script": provision_script,
-        "cleanup_script": _IGMP_CLEANUP_SCRIPT,
-        "artifact_root": "target/probe/target-services/igmp",
-        "deterministic": True,
-        "dry_run_safe": True,
-        "live_guard": "LIBCRAFTER_PROBE_LAB_TARGET=1",
-    }
-
-
-def deterministic_igmp_group(profile: str, seed: int, sequence: int) -> str:
-    digest = deterministic_bytes("igmp-group", profile, seed, sequence)
-    host = 1 + digest[0] % 254
-    return f"233.252.0.{host}"
-
-
-def deterministic_igmp_source_list(profile: str, seed: int, sequence: int) -> list[str]:
-    digest = deterministic_bytes("igmp-source-list", profile, seed, sequence)
-    first = 1 + digest[0] % 254
-    second = 1 + digest[1] % 254
-    return [f"192.0.2.{first}", f"198.51.100.{second}"]
 
 
 def _ipsec_probe_plan(
@@ -761,10 +486,10 @@ _LEGACY_PLAN_BUILDERS: dict[str, PlanBuilder] = {
     # The RIP/RIPng plan builders now live in ``protocols/rip.py`` and reach
     # ``PLAN_BUILDERS`` through the registry merge below (re-imported into this
     # module above for identity/back-compat).
-    "igmp-membership-query-observation": _igmp_probe_plan,
-    "igmp-v2-membership-report-emission": _igmp_probe_plan,
-    "igmp-v2-leave-group-emission": _igmp_probe_plan,
-    "igmp-v3-source-list-report": _igmp_probe_plan,
+    # The IGMP plan builder (``_igmp_probe_plan``, serving all four IGMP cases)
+    # now lives in ``protocols/igmp.py`` and reaches ``PLAN_BUILDERS`` through the
+    # registry merge below (re-imported into this module above for
+    # identity/back-compat).
     "esp-transport-echo": _ipsec_probe_plan,
     "esp-tunnel-echo": _ipsec_probe_plan,
     "ah-transport-verify": _ipsec_probe_plan,
@@ -797,11 +522,8 @@ _LEGACY_PLANNED_ONLY_REGISTERED_CASES: frozenset[str] = frozenset(
         # (``protocols/bgp.py``) and unioned into ``PLANNED_ONLY_REGISTERED_CASES``
         # below; the ``rip-update-v2`` / ``ripng-update`` planned-only cases are
         # contributed by the RIP plugin (``protocols/rip.py``) and unioned in the
-        # same way.
-        "igmp-membership-query-observation",
-        "igmp-v2-membership-report-emission",
-        "igmp-v2-leave-group-emission",
-        "igmp-v3-source-list-report",
+        # same way. The four IGMP planned-only cases are contributed by the IGMP
+        # plugin (``protocols/igmp.py``) and unioned in the same way.
     }
 )
